feat(pkg/lint): removed chart and dir name match
As part of the removal of the chart name and chart directory
name matching constraint, this updates the linting rules to no
longer error on a mismatch. The requirement that there be a Chart
name, covered by validateChartName rule, also covers the removed
test case in TestValidateChartNameDirMatch where the Chart name is
missing.

Closes #1979